django_places_autocomplete/requirements.txt
Django>=4.2
python-dotenv
psycopg2-binary